REFACTOR: Put all Python parts in the server directory
