--- config/docstrings.mustache
@@ -1,25 +1,0 @@
-{{summaryPlaceholder}}
-
-{{#argsExist}}
-{{#args}}
-- **{{var}}** ({{typePlaceholder}}): {{descriptionPlaceholder}}
-{{/args}}
-{{/argsExist}}
-
-{{#kwargsExist}}
-{{#kwargs}}
-- **{{var}}** ({{typePlaceholder}}): {{descriptionPlaceholder}} (default: {{&default}})
-{{/kwargs}}
-{{/kwargsExist}}
-
-{{#returnsExist}}
-{{#returns}}
-- **Returns** {{typePlaceholder}}: {{descriptionPlaceholder}} 
-{{/returns}} 
-{{/returnsExist}}
-
-{{#exceptionsExist}}
-{{#exceptions}}
-- **{{type}}**: {{descriptionPlaceholder}}
-{{/exceptions}}
-{{/exceptionsExist}}
