+++ server/api/routers/sizes.py
@@ -4,7 +4,7 @@ from sqlalchemy.orm import Session
 from sqlalchemy.exc import IntegrityError
 
 from server.api import get_db
-from config.logger_config import logger
+from server.config.logger_config import logger
 from server.api.auth.security import oauth2_scheme
 from server.models import Size
 from server.api.models.size import SizeModel, SizePostBase, SizePutBase, SizeResponse
