+++ server/api/utils/utilities.py
@@ -1,10 +1,9 @@
 from typing import Any
 
 from fastapi import HTTPException
-from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-from config.logger_config import logger
+from server.config.logger_config import logger
 
 
 class Utilities:
