+++ server/tests/test_auth.py
