+++ server/tests/test_classes.py
