+++ server/tests/test_damage_types.py
