+++ server/tests/test_non_player_characters.py
