+++ server/tests/test_parties.py
