+++ server/tests/test_types.py
