+++ server/tests/test_users.py
